Testing/board/scripts/requirements.txt
numpy == 1.26.4
colorama  == 0.4.6
opencv-python == 4.9.0.80
pillow == 10.2.0
